requirements.txt
pandas
fuzzywuzzy
biopython==1.72
